debug: udp send

+++ autopilot_udp_interface.cpp
@@ -573,7 +573,7 @@ write_mavlink_message(mavlink_message_t &message)
     // Write buffer to UDP socket
     //_write_port(buf,len);
     cout << "dest Address: " <<destAddr<<endl;
-    udp_socket->sendTo(buf, (int)len, destAddr, destPort);
+    udp_socket->sendTo(buf, len, destAddr, destPort);
     
     return len;
 }
